+++ lib/msal-browser/src/error/BrowserConfigurationAuthError.ts
@@ -31,6 +31,14 @@ export const BrowserConfigurationAuthErrorMessage = {
         desc: "The object passed for the callback was invalid. " +
           "More information is available here: https://github.com/AzureAD/microsoft-authentication-library-for-js/wiki/MSAL-basics."
     },
+    stubPcaInstanceCalled: {
+        code: "stubbed_public_client_application_called",
+        desc: "Stub instance of Public Client Application was called. If using msal-react, please ensure context is not used without a provider."
+    },
+    inMemRedirectUnavailable: {
+        code: "in_mem_redirect_unavailable",
+        desc: "Redirect cannot be supported. In-memory storage was selected and storeAuthStateInCookie=false, which would cause the library to be unable to handle the incoming hash. If you would like to use the redirect API, please use session/localStorage or set storeAuthStateInCookie=true."
+    }
 };
 
 /**
@@ -84,5 +92,20 @@ export class BrowserConfigurationAuthError extends AuthError {
     static createRedirectCallbacksNotSetError(): BrowserConfigurationAuthError {
         return new BrowserConfigurationAuthError(BrowserConfigurationAuthErrorMessage.noRedirectCallbacksSet.code, 
             BrowserConfigurationAuthErrorMessage.noRedirectCallbacksSet.desc);
+    }
+
+    /**
+     * Creates error thrown when the stub instance of PublicClientApplication is called.
+     */
+    static createStubPcaInstanceCalledError(): BrowserConfigurationAuthError {
+        return new BrowserConfigurationAuthError(BrowserConfigurationAuthErrorMessage.stubPcaInstanceCalled.code,
+            BrowserConfigurationAuthErrorMessage.stubPcaInstanceCalled.desc);
+    }
+
+    /*
+     * Create an error thrown when in-memory storage is used and storeAuthStateInCookie=false.
+     */
+    static createInMemoryRedirectUnavailableError(): BrowserConfigurationAuthError {
+        return new BrowserConfigurationAuthError(BrowserConfigurationAuthErrorMessage.inMemRedirectUnavailable.code, BrowserConfigurationAuthErrorMessage.inMemRedirectUnavailable.desc);
     }
 }