+++ lib/msal-core/src/utils/StringUtils.ts
@@ -3,6 +3,8 @@
  * Licensed under the MIT License.
  */
 
+import { AccessTokenKey } from "../cache/AccessTokenKey";
+
 /**
  * @hidden
  */
@@ -21,18 +23,18 @@ export class StringUtils {
      *
      * @param str
      */
-    static isValidJson(str: string): boolean {
+    static validateAndParseJsonCacheKey(str: string): AccessTokenKey {
         try {
-            const parsedValue = JSON.parse(str);
+            const parsedKey = JSON.parse(str);
             /**
              * There are edge cases in which JSON.parse will successfully parse a non-valid JSON object 
              * (e.g. JSON.parse will parse an escaped string into an unescaped string), so adding a type check
              * of the parsed value is necessary in order to be certain that the string represents a valid JSON object.
              *
              */
-            return (parsedValue && typeof parsedValue === "object");
+            return (parsedKey && typeof parsedKey === "object") ? parsedKey : null;
         } catch (error) {
-            return false;
+            return null;
         }
     }
 }