+++ lib/msal-browser/changelog.md
@@ -1,8 +1,55 @@
 # Change Log - @azure/msal-browser
 
-This log was last generated on Mon, 26 Oct 2020 21:00:29 GMT and should not be manually modified.
+This log was last generated on Tue, 10 Nov 2020 01:48:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.6.1
+
+Tue, 10 Nov 2020 01:48:44 GMT
+
+### Patches
+
+- Export stubbed PCA instance (#2540) (thomas.norling@microsoft.com)
+
+## 2.6.0
+
+Sat, 07 Nov 2020 01:50:14 GMT
+
+### Minor changes
+
+- Fixing a bug and adding `localAccountId` in AccountInfo interface (#2516) (sameera.gajjarapu@microsoft.com)
+
+## 2.5.2
+
+Mon, 02 Nov 2020 23:33:39 GMT
+
+### Patches
+
+- Fix JSON.parse issue and cache value validation (#2527) (prkanher@microsoft.com)
+
+## 2.5.1
+
+Fri, 30 Oct 2020 00:52:19 GMT
+
+### Patches
+
+- Restore string to cacheLocation type (#2523) (janutter@microsoft.com)
+
+## 2.5.0
+
+Thu, 29 Oct 2020 20:36:48 GMT
+
+### Minor changes
+
+- Add getLogger and setLogger to msal-browser (#2513) (joarroyo@microsoft.com)
+- Adding memory storage option for cache location (#2481) (prkanher@microsoft.com)
+
+### Patches
+
+- Add handleRedirect End Event (#2518) (thomas.norling@microsoft.com)
+- Ensure history.replaceState is a function (janutter@microsoft.com)
+- Allow hash to be passed into handleRedirectPromise, reset non-msal after processing (janutter@microsoft.com)
 
 ## 2.4.1